PhantomX works with Arbotix-Pro or mine...
Servo code updated to allow other servo controllers other than USB2AX

Add parameters to allow you to define, the ID of the servo controller,
plus what Baud num you wish to use for the controller and if the
controller needs for you to set a register on the device to turn the
power on to the servos.

This helps support CM-730/Arbotix Pro/ My Teensy based Arbotix Pro.

Warning the Phantomx.yaml is updated to my settings, which has the my
TTL connection to the Teensy running at 2mbs.