WIP - Servo Driver add Interpolation
Start of work to update the servo driver code to use Servo
Interpolation.
That is all of the servos in the move should move such that they all
arrive at the new location at about the same time.  That is if two
servos are moving and one is moving 10 units and the other is moving 2,
The released code would have the move take 10 iterations, where the
first servo would move 1 unit per iteration and the 2nd would move it's
two units on the first two iterations.  Change is that the 2nd should
move one unit on about iteration 5 and the 2nd on iteration 10...

New code is also first step to make the calls to the servo driver have a
time value associated with it, such that we can say the move should take
n MS and the driver will figure out how many iterations this should
be... Currently hard coded time...