Cleanup support for different controllers
Cleaned up the support for the different controllers.  Instead of using
my fixed
/dev/ttyDXL then fall back to /dev/ACM0 then back to /dev/ttyUSB0, I
know allow the yaml file to specify the device path to use.  Likewise I
changed the ability to pass in Baud index to instead pass in Baud.
Right now it uses a helper function to map baud to the predefined baud
defines like:
B1000000.

If the YAML file does not have the defines, the code will fall back to:
/dev/ttyACM0 at 1000000 as to be compatible with existing setups.

Tested on PhantomX, with my own Teensy 3.2 board connected with hardware
serial port.  Also tested with default of USB2AX at 1000000